requirements.txt
h5py>=3.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
cartopy>=0.21.0
Flask>=2.3.0
requests>=2.31.0
